refactor: MemorialHouseName UseCase 구현
