+++ MemorialHouse/MHDomain/MHDomain/UseCase/Interface/MemorialHouseNameUseCase.swift
@@ -1,0 +1,7 @@
+public protocol CreateMemorialHouseNameUseCase: Sendable {
+    func execute(with name: String) async throws
+}
+
+public protocol FetchMemorialHouseNameUseCase: Sendable {
+    func execute() async throws -> String
+}
